// src/types/AllTypes.ts
export interface Job {
  id: string;
  jobId: string;
  title: string;
  experience: string;
  deadline: string;
  location: string;
  salaryType: string;
  salaryRange: string;
  noOfApplicants: number;
  skills: string[];
  features: {
    featureTitle: string;
    paragraph?: string;
    point?: string[];
  }[];
  userId: string;
  companyId: string;
  jobType: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  user?: {
    id: string;
    firstName: string;
    lastName: string;
    fullName: string;
    email: string;
    profilePic: string;
    role: string;
    isVerified?: boolean;
  };
  company?: {
    companyName: string;
    logo: string;
    industryType: string;
    email: string;
    phoneNumber: string;
  };
}